[semi-supervised] db_sampler관련 함수 정리
